+++ src/enumerate.rs
@@ -1,7 +1,7 @@
 use std::{iter::Zip, ops::Range, slice::Iter, str::Chars};
 
 /// `enumerate` を実装するためのトレイト。
-pub trait Enumerator<'a> {
+pub trait Iterable<'a> {
     type Item;
 
     /// インデックス付きでイテレートします。
@@ -10,13 +10,13 @@ pub trait Enumerator<'a> {
     fn enumerate(&'a self, start: i32) -> Zip<Range<i32>, Self::Item>;
 }
 
-impl<'a> Enumerator<'a> for str {
+impl<'a> Iterable<'a> for str {
     type Item = Chars<'a>;
 
     /// # Examples
     ///
     /// ```
-    /// use rspy::Enumerator;
+    /// use rspy::Iterable;
     ///
     /// let mut index_vec = vec![];
     /// let mut char_vec = vec![];
@@ -30,31 +30,31 @@ impl<'a> Enumerator<'a> for str {
     /// assert_eq!(char_vec, ['a', 'b', 'c', 'd', 'e']);
     /// ```
     fn enumerate(&'a self, start: i32) -> Zip<Range<i32>, Self::Item> {
-        (start..self.len() as i32).zip(self.chars())
+        (start..start + self.len() as i32).zip(self.chars())
     }
 }
 
-impl<'a> Enumerator<'a> for String {
+impl<'a> Iterable<'a> for String {
     type Item = Chars<'a>;
 
     fn enumerate(&'a self, start: i32) -> Zip<Range<i32>, Self::Item> {
-        (start..self.len() as i32).zip(self.chars())
+        (start..start + self.len() as i32).zip(self.chars())
     }
 }
 
-impl<'a, T: 'a> Enumerator<'a> for Vec<T> {
+impl<'a, T: 'a> Iterable<'a> for Vec<T> {
     type Item = Iter<'a, T>;
 
     fn enumerate(&'a self, start: i32) -> Zip<Range<i32>, Self::Item> {
-        (start..self.len() as i32).zip(self.iter())
+        (start..start + self.len() as i32).zip(self.iter())
     }
 }
 
-impl<'a, T: 'a, const N: usize> Enumerator<'a> for [T; N] {
+impl<'a, T: 'a, const N: usize> Iterable<'a> for [T; N] {
     type Item = Iter<'a, T>;
 
     fn enumerate(&'a self, start: i32) -> Zip<Range<i32>, Self::Item> {
-        (start..self.len() as i32).zip(self.iter())
+        (start..start + self.len() as i32).zip(self.iter())
     }
 }
 
@@ -79,6 +79,20 @@ mod tests {
     }
 
     #[test]
+    fn str_enumerate_works2() {
+        let mut index_vec = vec![];
+        let mut char_vec = vec![];
+
+        for (i, ch) in "abcde".enumerate(3) {
+            index_vec.push(i);
+            char_vec.push(ch);
+        }
+
+        assert_eq!(index_vec, [3, 4, 5, 6, 7]);
+        assert_eq!(char_vec, ['a', 'b', 'c', 'd', 'e']);
+    }
+
+    #[test]
     fn string_enumerate_works() {
         let mut index_vec = vec![];
         let mut char_vec = vec![];
@@ -91,6 +105,22 @@ mod tests {
         }
 
         assert_eq!(index_vec, [-3, -2, -1, 0, 1]);
+        assert_eq!(char_vec, ['a', 'b', 'c', 'd', 'e']);
+    }
+
+    #[test]
+    fn string_enumerate_works2() {
+        let mut index_vec = vec![];
+        let mut char_vec = vec![];
+
+        let s = "abcde".to_string();
+
+        for (i, ch) in s.enumerate(3) {
+            index_vec.push(i);
+            char_vec.push(ch);
+        }
+
+        assert_eq!(index_vec, [3, 4, 5, 6, 7]);
         assert_eq!(char_vec, ['a', 'b', 'c', 'd', 'e']);
     }
 
@@ -113,6 +143,24 @@ mod tests {
     }
 
     #[test]
+    fn vec_enumerate_works2() {
+        let mut index_vec = vec![];
+        let mut int_vec = vec![];
+
+        let vec = vec![100, -100, 20, 50, -1000];
+
+        for (i, v) in vec.enumerate(3) {
+            index_vec.push(i);
+            int_vec.push(v);
+        }
+
+        assert_eq!(index_vec, [3, 4, 5, 6, 7]);
+        for (i, v) in vec.iter().enumerate() {
+            assert_eq!(int_vec[i], v);
+        }
+    }
+
+    #[test]
     fn arr_enumerate_works() {
         let mut index_vec = vec![];
         let mut int_vec = vec![];
@@ -125,6 +173,24 @@ mod tests {
         }
 
         assert_eq!(index_vec, [-3, -2, -1, 0, 1]);
+        for (i, v) in arr.iter().enumerate() {
+            assert_eq!(int_vec[i], v);
+        }
+    }
+
+    #[test]
+    fn arr_enumerate_works2() {
+        let mut index_vec = vec![];
+        let mut int_vec = vec![];
+
+        let arr = [100, -100, 20, 50, -1000];
+
+        for (i, v) in arr.enumerate(3) {
+            index_vec.push(i);
+            int_vec.push(v);
+        }
+
+        assert_eq!(index_vec, [3, 4, 5, 6, 7]);
         for (i, v) in arr.iter().enumerate() {
             assert_eq!(int_vec[i], v);
         }