+++ src/lib.rs
@@ -2,8 +2,9 @@
 
 mod enumerate;
 mod input;
+pub mod pathlib;
 pub mod string;
 pub mod time;
 
-pub use crate::enumerate::Enumerator;
+pub use crate::enumerate::Iterable;
 pub use crate::input::input;